requirements.txt
PyMuPDF==1.24.10
pytesseract==0.3.13
Pillow==10.4.0
Flask==3.0.3


